moved functions in Merkle
