+++ Sources/BeaconChain/StateTransition.swift
@@ -9,7 +9,7 @@ class StateTransition {
         state.latestBlockRoots[Int((state.slot - 1) % LATEST_BLOCK_ROOTS_LENGTH)] = previousBlockRoot
 
         if state.slot % LATEST_BLOCK_ROOTS_LENGTH == 0 {
-            state.batchedBlockRoots.append(BeaconChain.merkleRoot(values: state.latestBlockRoots))
+            state.batchedBlockRoots.append(Merkle.root(state.latestBlockRoots))
         }
     }
 }
@@ -224,7 +224,7 @@ extension StateTransition {
             let serializedDepositData = Data(count: 64) // @todo when we have SSZ
 
             assert(
-                verifyMerkleBranch(
+                Merkle.verifyBranch(
                     leaf: BeaconChain.hash(serializedDepositData),
                     branch: deposit.branch,
                     depth: Int(DEPOSIT_CONTRACT_TREE_DEPTH),
@@ -296,19 +296,6 @@ extension StateTransition {
             state.validatorBalances[Int(transfer.to)] += transfer.amount
             state.validatorBalances[Int(BeaconChain.getBeaconProposerIndex(state: state, slot: state.slot))] += transfer.fee
         }
-    }
-
-    static func verifyMerkleBranch(leaf: Bytes32, branch: [Bytes32], depth: Int, index: Int, root: Bytes32) -> Bool {
-        var value = leaf
-        for i in 0..<depth {
-            if index / (2 ** i) % 2 == 1 {
-                value = BeaconChain.hash(branch[i] + value)
-            } else {
-                value = BeaconChain.hash(value + branch[i])
-            }
-        }
-
-        return value == root
     }
 }
